+++ src/assertions.ts
@@ -48,7 +48,8 @@ import {
   type TestCase,
   isGradingResult,
 } from './types';
-import { extractJsonObjects, isJavascriptFile } from './util';
+import { isJavascriptFile } from './util';
+import { extractJsonObjects } from './util/json';
 import { getNunjucksEngine } from './util/templates';
 import { transform } from './util/transform';
