+++ src/util/index.ts
@@ -1008,36 +1008,6 @@ export function renderVarsInObject<T>(obj: T, vars?: Record<string, string | obj
   return obj;
 }
 
-export function extractJsonObjects(str: string): object[] {
-  // This will extract all json objects from a string
-
-  const jsonObjects = [];
-  let openBracket = str.indexOf('{');
-  let closeBracket = str.indexOf('}', openBracket);
-  // Iterate over the string until we find a valid JSON-like pattern
-  // Iterate over all trailing } until the contents parse as json
-  while (openBracket !== -1) {
-    const jsonStr = str.slice(openBracket, closeBracket + 1);
-    try {
-      jsonObjects.push(JSON.parse(jsonStr));
-      // This is a valid JSON object, so start looking for
-      // an opening bracket after the last closing bracket
-      openBracket = str.indexOf('{', closeBracket + 1);
-      closeBracket = str.indexOf('}', openBracket);
-    } catch (err) {
-      // Not a valid object, move on to the next closing bracket
-      closeBracket = str.indexOf('}', closeBracket + 1);
-      while (closeBracket === -1) {
-        // No closing brackets made a valid json object, so
-        // start looking with the next opening bracket
-        openBracket = str.indexOf('{', openBracket + 1);
-        closeBracket = str.indexOf('}', openBracket);
-      }
-    }
-  }
-  return jsonObjects;
-}
-
 /**
  * Parses a file path or glob pattern to extract function names and file extensions.
  * Function names can be specified in the filename like this:
