+++ test/util.json.test.ts
@@ -1,5 +1,5 @@
 import dedent from 'dedent';
-import { isValidJson, safeJsonStringify, orderKeys } from '../src/util/json';
+import { extractJsonObjects, isValidJson, safeJsonStringify, orderKeys } from '../src/util/json';
 
 describe('json utilities', () => {
   describe('isValidJson', () => {
@@ -83,6 +83,38 @@ describe('json utilities', () => {
     it('preserves non-circular nested structures', () => {
       const nested = { a: { b: { c: 1 } }, d: [1, 2, { e: 3 }] };
       expect(JSON.parse(safeJsonStringify(nested))).toEqual(nested);
+    });
+  });
+
+  describe('extractJsonObjects', () => {
+    it('should extract a single JSON object from a string', () => {
+      const input = '{"key": "value"}';
+      const expectedOutput = [{ key: 'value' }];
+      expect(extractJsonObjects(input)).toEqual(expectedOutput);
+    });
+
+    it('should extract multiple JSON objects from a string', () => {
+      const input = 'yolo {"key1": "value1"} some text {"key2": "value2"} fomo';
+      const expectedOutput = [{ key1: 'value1' }, { key2: 'value2' }];
+      expect(extractJsonObjects(input)).toEqual(expectedOutput);
+    });
+
+    it('should return an empty array if no JSON objects are found', () => {
+      const input = 'no json here';
+      const expectedOutput: any[] = [];
+      expect(extractJsonObjects(input)).toEqual(expectedOutput);
+    });
+
+    it('should handle nested JSON objects', () => {
+      const input = 'wassup {"outer": {"inner": "value"}, "foo": [1,2,3,4]}';
+      const expectedOutput = [{ outer: { inner: 'value' }, foo: [1, 2, 3, 4] }];
+      expect(extractJsonObjects(input)).toEqual(expectedOutput);
+    });
+
+    it('should handle invalid JSON gracefully', () => {
+      const input = '{"key": "value" some text {"key2": "value2"}';
+      const expectedOutput = [{ key2: 'value2' }];
+      expect(extractJsonObjects(input)).toEqual(expectedOutput);
     });
   });
