+++ test/util.test.ts
@@ -6,7 +6,6 @@ import * as googleSheets from '../src/googleSheets';
 import type { ApiProvider, EvaluateResult, EvaluateTable, TestCase } from '../src/types';
 import {
   maybeLoadFromExternalFile,
-  extractJsonObjects,
   isJavascriptFile,
   parsePathOrGlob,
   providerToIdentifier,
@@ -648,38 +647,6 @@ describe('util', () => {
       };
 
       expect(resultIsForTestCase(result, nonMatchTestCase)).toBe(false);
-    });
-  });
-
-  describe('extractJsonObjects', () => {
-    it('should extract a single JSON object from a string', () => {
-      const input = '{"key": "value"}';
-      const expectedOutput = [{ key: 'value' }];
-      expect(extractJsonObjects(input)).toEqual(expectedOutput);
-    });
-
-    it('should extract multiple JSON objects from a string', () => {
-      const input = 'yolo {"key1": "value1"} some text {"key2": "value2"} fomo';
-      const expectedOutput = [{ key1: 'value1' }, { key2: 'value2' }];
-      expect(extractJsonObjects(input)).toEqual(expectedOutput);
-    });
-
-    it('should return an empty array if no JSON objects are found', () => {
-      const input = 'no json here';
-      const expectedOutput: any[] = [];
-      expect(extractJsonObjects(input)).toEqual(expectedOutput);
-    });
-
-    it('should handle nested JSON objects', () => {
-      const input = 'wassup {"outer": {"inner": "value"}, "foo": [1,2,3,4]}';
-      const expectedOutput = [{ outer: { inner: 'value' }, foo: [1, 2, 3, 4] }];
-      expect(extractJsonObjects(input)).toEqual(expectedOutput);
-    });
-
-    it('should handle invalid JSON gracefully', () => {
-      const input = '{"key": "value" some text {"key2": "value2"}';
-      const expectedOutput = [{ key2: 'value2' }];
-      expect(extractJsonObjects(input)).toEqual(expectedOutput);
     });
   });
